backend/Tutorium/pyproject.toml
[tool.poetry]
name = "tutorium"
version = "0.1.0"
description = ""
authors = ["Baris Ogun Yoruk <barisoyoruk@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
start = "tutorium.main:start"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
mysqlclient = "^2.1.1"
sqlalchemy = "^2.0.13"
pyopenssl = "^23.1.1"
pydantic = "^1.10.7"
alembic = "^1.10.4"
jwt = "^1.3.1"
python-jose = "^3.3.0"
flask = "^2.3.2"
flask-cors = "^3.0.10"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
